исправлено задание 4

+++ main.cpp
@@ -9,8 +9,26 @@
 
 using namespace std;
 
+/**
+ * @brief Вводит элементы вектора с клавиатуры
+ * @param vec Ссылка на вектор для заполнения
+ * @param name Имя вектора
+ */
 void inputVector(vector<int>& vec, const string& name);
+
+
+/**
+ * @brief Выводит содержимое вектора на экран
+ * @param vec Константная ссылка на вектор
+ */
 void printVector(const vector<int>& vec);
+
+/**
+ * @brief Подсчитывает количество векторов Vi, содержащих все элементы V0
+ * @param V0 Исходный вектор
+ * @param vectors Набор векторов V1, ..., VN
+ * @return Количество подходящих векторов
+ */
 int countMatchingVectors(const vector<int>& V0, const vector<vector<int>>& vectors);
 
 /**
@@ -34,14 +52,14 @@ int main() {
     }
 
     vectors.resize(N);
-    for (int i = 0; i < N; i++) {
+    for (size_t i = 0; i < N; i++) {
         inputVector(vectors[i], "V" + to_string(i + 1));
     }
     
     cout << "\nВектор V0: ";
     printVector(V0);
     
-    for (int i = 0; i < N; i++) {
+    for (size_t i = 0; i < N; i++) {
         cout << "Вектор V" << i + 1 << ": ";
         printVector(vectors[i]);
     }
@@ -54,11 +72,6 @@ int main() {
     return 0;
 }
 
-/**
- * @brief Вводит элементы вектора с клавиатуры
- * @param vec Ссылка на вектор для заполнения
- * @param name Имя вектора
- */
 void inputVector(vector<int>& vec, const string& name) {
     cout << "Введите вектор " << name << ": ";
     
@@ -68,10 +81,6 @@ void inputVector(vector<int>& vec, const string& name) {
     vec.assign(istream_iterator<int>(iss), istream_iterator<int>());
 }
 
-/**
- * @brief Выводит содержимое вектора на экран
- * @param vec Константная ссылка на вектор
- */
 void printVector(const vector<int>& vec) {
     if (vec.empty()) {
         cout << "пуст" << endl;
@@ -84,12 +93,6 @@ void printVector(const vector<int>& vec) {
     cout << endl;
 }
 
-/**
- * @brief Подсчитывает количество векторов Vi, содержащих все элементы V0
- * @param V0 Исходный вектор
- * @param vectors Набор векторов V1, ..., VN
- * @return Количество подходящих векторов
- */
 int countMatchingVectors(const vector<int>& V0, const vector<vector<int>>& vectors) {
     int count = 0;
